Add onCancel support to ForwardStreamController
The previous version of this patch actively checks if the connection is still in a collection, and reasoning about memory and possible retain cycles, especially w/a GC, seems difficult.

Here, we have native adapt a pattern like web's Channel implementation.

On web, an onCancel handler is passed to the StreamController to handle resource cleanup.

In this patch, ForwardStreamController now takes an onCancel handler.

Tests continue to pass with this implementation, and the design didn't need to change, as it is checking to make sure activeConnections is cleared after a response, which would be the net effect of any patch.